+++ bbs/lpfunc.cpp
@@ -138,7 +138,7 @@ int listfiles_plus_function(int type) {
   int max_lines = calc_max_lines();
 
   bool all_done = false;
-  for (size_t this_dir = 0; (this_dir < a()->directories.size()) && (!hangup) && (a()->udir[this_dir].subnum != -1)
+  for (uint16_t this_dir = 0; (this_dir < a()->directories.size()) && (!hangup) && (a()->udir[this_dir].subnum != -1)
        && !all_done; this_dir++) {
     int also_this_dir = a()->udir[this_dir].subnum;
     bool scan_dir = false;
